+++ src/ServiceManagement/Compute/Commands.ServiceManagement/IaaS/Disks/RemoveAzureDataDisk.cs
@@ -17,12 +17,13 @@ using System.Collections.ObjectModel;
 using System.Globalization;
 using System.Linq;
 using System.Management.Automation;
+using Microsoft.WindowsAzure.Commands.ServiceManagement.Common;
 using Microsoft.WindowsAzure.Commands.ServiceManagement.Model;
 using Microsoft.WindowsAzure.Commands.ServiceManagement.Properties;
 
 namespace Microsoft.WindowsAzure.Commands.ServiceManagement.IaaS
 {
-    [Cmdlet(VerbsCommon.Remove, "AzureDataDisk"), OutputType(typeof(IPersistentVM))]
+    [Cmdlet(VerbsCommon.Remove, ProfileNouns.DataDisk), OutputType(typeof(IPersistentVM))]
     public class RemoveAzureDataDiskCommand : VirtualMachineConfigurationCmdletBase
     {
         [Parameter(Position = 0, Mandatory = true, HelpMessage = "Numerical value that defines the slot where the data drive is mounted in the virtual machine.")]